#1446 Move undo/redo handling logic to a separate class
